// modules/cinema/frontend/src/lib/ws-client-msg.gen.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type ClientMsg =
  | {
      episode?: number | null;
      file_idx: number;
      info_hash: string;
      media_type: string;
      poster_path?: string | null;
      season?: number | null;
      title: string;
      tmdb_id: number;
      type: "set_content";
      [k: string]: unknown;
    }
  | {
      type: "navigate";
      url: string;
      [k: string]: unknown;
    }
  | {
      type: "loaded";
      [k: string]: unknown;
    }
  | {
      type: "ready";
      [k: string]: unknown;
    }
  | {
      type: "unready";
      [k: string]: unknown;
    }
  | {
      position: number;
      type: "play";
      [k: string]: unknown;
    }
  | {
      position: number;
      type: "pause";
      [k: string]: unknown;
    }
  | {
      position: number;
      type: "seek";
      [k: string]: unknown;
    }
  | {
      type: "ping";
      [k: string]: unknown;
    };