// modules/cinema/frontend/src/lib/ws-server-msg.gen.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type ServerMsg =
  | {
      code: string;
      type: "room_created";
      [k: string]: unknown;
    }
  | {
      id: number;
      role: Role;
      type: "welcome";
      [k: string]: unknown;
    }
  | {
      count: number;
      id: number;
      type: "participant_joined";
      [k: string]: unknown;
    }
  | {
      count: number;
      id: number;
      type: "participant_left";
      [k: string]: unknown;
    }
  | {
      reason: string;
      type: "room_closed";
      [k: string]: unknown;
    }
  | {
      episode?: number | null;
      file_idx: number;
      info_hash: string;
      media_type: string;
      poster_path?: string | null;
      season?: number | null;
      title: string;
      tmdb_id: number;
      type: "content_set";
      [k: string]: unknown;
    }
  | {
      type: "navigate_sync";
      url: string;
      [k: string]: unknown;
    }
  | {
      ready: number[];
      total: number;
      type: "ready_state";
      [k: string]: unknown;
    }
  | {
      playing: boolean;
      position: number;
      server_time: number;
      type: "sync";
      [k: string]: unknown;
    }
  | {
      type: "pong";
      [k: string]: unknown;
    }
  | {
      message: string;
      type: "error";
      [k: string]: unknown;
    };
export type Role = "host" | "guest";